+++ lib/EasyPost/Service/BetaRateService.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace EasyPost\Service;
+
+use EasyPost\Http\Requestor;
+use EasyPost\Util\InternalUtil;
+
+/**
+ * Rate service containing all the logic to make API calls.
+ */
+class BetaRateService extends BaseService
+{
+    /**
+     * Retreive stateless rates.
+     *
+     * @param mixed $params
+     * @return mixed
+     */
+    public function retrieveStatelessRates($params)
+    {
+        $wrappedParams = [
+            'shipment' => $params,
+        ];
+
+        $response = Requestor::request($this->client, 'post', '/rates', $wrappedParams, true)['rates'];
+
+        foreach ($response as &$key) {
+            $key['object'] = 'StatelessRate';
+            // Convert object from `rate` to `stateless` in order to deserialize the response to Stateless Object
+        }
+
+        return InternalUtil::convertToEasyPostObject($this->client, $response);
+    }
+}