+++ lib/EasyPost/Util/InternalUtil.php
@@ -28,6 +28,7 @@ use EasyPost\Refund;
 use EasyPost\Report;
 use EasyPost\ScanForm;
 use EasyPost\Shipment;
+use EasyPost\StatelessRate;
 use EasyPost\Tracker;
 use EasyPost\TrackingDetail;
 use EasyPost\TrackingLocation;
@@ -99,6 +100,7 @@ abstract class InternalUtil
             'Shipment'              => Shipment::class,
             'ShipmentInvoiceReport' => Report::class,
             'ShipmentReport'        => Report::class,
+            'StatelessRate'         => StatelessRate::class,
             'TaxIdentifier'         => TaxIdentifier::class,
             'Tracker'               => Tracker::class,
             'TrackerReport'         => Report::class,